+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppSdkException.kt
@@ -15,6 +15,20 @@ open class MiniAppSdkException(message: String, cause: Throwable?) : Exception(m
 }
 
 /**
+ * Exception which is thrown when the server returns no published
+ * versions for the provided mini app ID.
+ */
+class MiniAppHasNoPublishedVersionException(appId: String) :
+    MiniAppSdkException("Server returned no published version info for the provided Mini App Id: $appId")
+
+/**
+ * Exception which is thrown when the provided mini app ID
+ * does not exist on the server.
+ */
+class MiniAppNotFoundException(serverMessage: String) :
+    MiniAppSdkException("$serverMessage: Server returned no mini app for the provided Mini App ID.")
+
+/**
  * Exception which is thrown when HostApp doesn't implement requestCustomPermissions interface.
  */
 internal class CustomPermissionsNotImplementedException :