+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.X.X (In progress)
 **SDK**
 - **Feature:** Provide the access token to miniapp.
-- **Feature:** Added default UI support for managing custom permissions in case `requestCustomPermissions` hasn't been implemented in Host App. 
+- **Feature:** Added default UI support for managing custom permissions in case `requestCustomPermissions` hasn't been implemented in Host App.
+- **Feature:** Added `MiniAppHasNoPublishedVersionException` and `MiniAppNotFoundException` exception types to `MiniApp.create` and `MiniApp.fetchInfo`.
 
 ### 2.4.0 (2020-10-30)
 **SDK**