+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - **Feature:** Mini App can call media execution play/pause programmatically.
 - **Change:** Added the default implementation for external link handler. Using [custom tab](https://developers.google.com/web/android/custom-tabs).
 
+**Sample App**
+- **Change:** Update setting of external webview.
+
 ### 2.4.0 (2020-10-30)
 **SDK**
 - **Feature:** Handle the screen orientation change request from miniapp.